+++ development/main.ts
@@ -11,6 +11,8 @@ import MultiSelect from '../tests/pages/MultiSelect'
 import Errors from "../tests/pages/Errors";
 import RadioGroupTests from "../tests/pages/RadioGroupTests";
 import EventHandlerTests from "../tests/pages/EventHandlerTests";
+import CheckboxTests from "../tests/pages/CheckboxTests";
+import RightToLeftTests from "../tests/pages/RightToLeft";
 
 import ReactDOM from 'react-dom'
 // @ts-ignore
@@ -30,6 +32,8 @@ const router = createRouter({
     { path: '/e2e/errors', component: Errors },
     { path: '/e2e/radiogroup', component: RadioGroupTests },
     { path: '/e2e/event-handlers', component: EventHandlerTests },
+    { path: '/e2e/checkbox', component: CheckboxTests },
+    { path: '/e2e/right-to-left', component: RightToLeftTests },
   ]
 })