+++ packages/core/src/hooks/use-required.ts
@@ -1,15 +1,14 @@
-import {Hook} from "@phormal/core/src/types/interfaces/Hook.interface";
+import {Hook, HookReturnValue} from "@phormal/core/src/types/interfaces/Hook.interface";
 import {FormField} from "@phormal/core/src/FormField";
+import {ValidationHelper} from "../util/validation-helper";
 
 const ERROR_NAME = 'required';
-export const useRequired: Hook = () => {
+export const useRequired = (): HookReturnValue => {
   return {
     validators: {
       checkRequired(field: FormField) {
         const isValid = !!field.getValue() || field.getValue() === '0'
-
-        if (!field.errors.includes(ERROR_NAME) && !isValid) field.errors.push(ERROR_NAME)
-        if (field.errors.includes(ERROR_NAME) && isValid) field.errors.splice(field.errors.indexOf(ERROR_NAME), 1)
+        ValidationHelper.pushOrSpliceError(field, ERROR_NAME, isValid)
       }
     },
 
@@ -19,7 +18,7 @@ export const useRequired: Hook = () => {
         ru: 'Этот ввод необходим',
         es: 'Esta entrada es necesaria',
         tr: 'Bu girdi gereklidir',
-        // fa: 'این ورودی مورد نیاز است',
+        fa: 'این ورودی مورد نیاز است',
         fr: 'Cette entrée est requise',
         de: 'Dieses Feld ist ein Pflichtfeld',
         ja: 'この入力は必須です',