+++ packages/use-valid-zip/src/index.ts
@@ -1,3 +1,0 @@
-import {useValidZip} from "./use-valid-zip";
-
-export default useValidZip;