+++ tests/components/code-element.ts
@@ -13,6 +13,7 @@ export default defineComponent({
   render() {
     return h(
       'code',
+      {id: 'code-element'},
       [
         h('pre', { style: { backgroundColor: '#000', color: 'yellow', padding: '1em' }, id: 'code' }, [JSON.stringify(this.code)])
       ]